support sending eventkit events in batches...
We have a specific Destination implementation which directly saves the Events to BQ (instead of sending via UDP to a daemon, which saves them to BQ).

But this endpoint doesn't have any queue.

In this patch, I copy pasted the queue logic from the UDP client and made a generic BatchQueue destination.

Change-Id: I25ad4c3c2c9a2a261985b9efab1cd12f9acec9cc